requirements.txt
torch
torchvision
torchaudio
transformers
peft
trl
bitsandbytes
accelerate
datasets
fastapi
uvicorn
python-multipart
pydantic
streamlit
pillow
python-dotenv
asyncio
langchain
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers
ollama
pypdf
